part2

+++ BOT/MovieAdminBot/app/models/model_user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 
 from app.db.db import Base
 
@@ -6,7 +6,7 @@ from app.db.db import Base
 class Users(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     first_name = Column(String(255))
     last_name = Column(String(255))
     user_id = Column(Integer, nullable=False, index=True)
